+++ VeritasEngine/BufferIndicies.cpp
@@ -1,7 +1,7 @@
 #include "BufferIndicies.h"
 
 
-VeritasEngine::BufferIndicies::BufferIndicies(unsigned int startIndex, unsigned int numberOfElements)
+VeritasEngine::BufferIndicies::BufferIndicies(size_t startIndex, size_t numberOfElements)
 	: StartIndex{ startIndex }, NumberOfElements{ numberOfElements }
 {
