+++ VeritasEngine_Windows/VertexBufferImpl.cpp
@@ -28,11 +28,11 @@ VeritasEngine::VertexBufferImpl::VertexBufferImpl()
 
 VeritasEngine::VertexBufferImpl::~VertexBufferImpl() = default;
 
-void VeritasEngine::VertexBufferImpl::SetData(unsigned int numOfElements, unsigned int sizeOfEachVertex, void* bufferData)
+void VeritasEngine::VertexBufferImpl::SetData(size_t numOfElements, size_t sizeOfEachVertex, void* bufferData)
 {
 	D3D11_BUFFER_DESC bufferDesc;
 	bufferDesc.Usage = D3D11_USAGE_DEFAULT;
-	bufferDesc.ByteWidth = sizeOfEachVertex * numOfElements;
+	bufferDesc.ByteWidth = static_cast<UINT>(sizeOfEachVertex * numOfElements);
 	bufferDesc.BindFlags = D3D11_BIND_VERTEX_BUFFER;
 	bufferDesc.CPUAccessFlags = 0;
 	bufferDesc.MiscFlags = 0;
