+++ VeritasGame/RotateObjectProcess.cpp
@@ -21,7 +21,7 @@ void RotateObjectProcess::OnUpdate(TimeDuration delta)
 	auto worldMatrix = VeritasEngine::SceneGraphProperties::WorldPosition.GetProperty(m_handle);
 
 	auto count = chrono::duration_cast<chrono::seconds>(delta).count();
-	auto rotationMatrix = VeritasEngine::MathHelpers::CreateRotationRollPitchYawMatrix(0.5 * count, 0.0, 0.0);
+	auto rotationMatrix = VeritasEngine::MathHelpers::CreateRotationRollPitchYawMatrix(static_cast<float>(0.5 * count), 0.0, 0.0);
 
 	auto result = *worldMatrix * rotationMatrix;
